fix: token state after refresh
