--- models/context.ts
@@ -1,6 +1,0 @@
-export type AuthContextType = {
-  isSignedIn: boolean;
-  token: string;
-  signIn: () => void;
-  signOut: () => void;
-};
